fix: rename depreciated frame to deprecated
