+++ src/components/three/assets/frame/deprecated/ImagePlane.tsx
