+++ src/components/three/assets/frame/deprecated/Picture_frame.tsx
